requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==3.25.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lit==15.0.7
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
slowapi==0.1.9
sympy==1.14.0
tornado==6.5.4
traitlets==5.14.3
triton==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.246
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2